requirements.txt
attrs==16.1.0
boto==2.43.0
certifi==2016.8.31
cffi==1.7.0
cryptography==1.5
cssselect==0.9.2
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10
django-bootstrap3==7.0.1
django-el-pagination==3.0.1
django-haystack==2.5.0
django-storages==1.5.1
elasticsearch==1.9.0
gunicorn==19.6.0
idna==2.1
lxml==3.6.4
parsel==1.0.3
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
PyDispatcher==2.0.5
pyelasticsearch==1.4
pyOpenSSL==16.1.0
queuelib==1.4.2
requests==2.11.1
Scrapy==1.1.2
scrapy-djangoitem==1.1.1
service-identity==16.0.0
simplejson==3.8.2
six==1.10.0
static3==0.7.0
Twisted==16.4.0
urllib3==1.17
w3lib==1.15.0
zope.interface==4.3.1
psycopg2>=2.6.1
